completed follow tests
