+++ src/test/java/ru/yandex/practicum/filmorate/DbTests/FollowDbTest.java
@@ -6,62 +6,143 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.jdbc.core.JdbcTemplate;
-import ru.yandex.practicum.filmorate.model.film.Director;
-import ru.yandex.practicum.filmorate.model.user.Follow;
+import ru.yandex.practicum.filmorate.model.user.User;
 import ru.yandex.practicum.filmorate.storage.follow.FollowDbStorage;
+import ru.yandex.practicum.filmorate.storage.user.UserDbStorage;
 
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertNull;
+import java.sql.Date;
+import java.util.ArrayList;
+import java.util.List;
+
+import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest
 @AutoConfigureTestDatabase
 @RequiredArgsConstructor(onConstructor_ = @Autowired)
 public class FollowDbTest {
-   /* private final JdbcTemplate jdbcTemplate;
+    private final JdbcTemplate jdbcTemplate;
     private final FollowDbStorage followDbStorage;
+    private final UserDbStorage userDbStorage;
 
-    private Follow getFollow() {
-        return Follow.builder().userId(1L).friendId(2L).build();
+    private User getUser() {
+        return User.builder().email("test@mail.ru").login("testLogin").name("testName").birthday(Date.valueOf("1946-08-20").toLocalDate()).build();
     }
 
-    private Follow getCommon() {
-        return Follow.builder().userId(3L).friendId(2L).build();
-    }
     @AfterEach
     void tearDown() {
         jdbcTemplate.update("DELETE FROM FOLLOW");
+        jdbcTemplate.update("DELETE FROM USERS");
+        jdbcTemplate.update("ALTER TABLE USERS ALTER COLUMN ID RESTART WITH 1");
     }
+
     @Test
     public void testAddFriend() {
-        Follow follow = getFollow();
-        followDbStorage.addFriend(follow.getUserId(), follow.getFriendId());
-        assertNotNull(followDbStorage.getAllFriends(follow.getUserId()), "Friends is empty");
+        userDbStorage.createUser(getUser());
+        User user = getUser();
+        user.setEmail("test1@mail.ru");
+        user.setLogin("testLogin1");
+        userDbStorage.createUser(user);
+        assertDoesNotThrow(() -> followDbStorage.addFriend(1L, 2L));
     }
 
     @Test
-    public void testDeleteFriend() {
-        Follow follow = getFollow();
-        followDbStorage.addFriend(follow.getUserId(), follow.getFriendId());
-        followDbStorage.deleteFriend(follow.getUserId(), follow.getFriendId());
-        assertNull(followDbStorage.getAllFriends(follow.getUserId()), "Friends is not empty after delete");
+    void testAddUnknownFriend() {
+        userDbStorage.createUser(getUser());
+        assertThrows(DataIntegrityViolationException.class, () -> followDbStorage.addFriend(1L, -1L), "Пользователь с id " + -1L + " не найден.");
     }
 
     @Test
-    public void testGetAllFriends() {
-        Follow follow = getFollow();
-        followDbStorage.addFriend(follow.getUserId(), follow.getFriendId());
-        assertNotNull(followDbStorage.getAllFriends(1L), "Friends is empty");
+    void testGetOneUserFriend() {
+        userDbStorage.createUser(getUser());
+        User user = getUser();
+        user.setEmail("test1@mail.ru");
+        user.setLogin("testLogin1");
+        User savedUser = userDbStorage.createUser(user);
+        followDbStorage.addFriend(1L, 2L);
+        assertEquals(List.of(savedUser), followDbStorage.getAllFriends(1L));
     }
 
     @Test
-    public void testGetCommonFriends() {
-        Follow follow = getFollow();
-        followDbStorage.addFriend(follow.getUserId(), follow.getFriendId());
-        Follow common = getCommon();
-        followDbStorage.addFriend(common.getUserId(), common.getFriendId());
+    void testGetTwoUserFriends() {
+        userDbStorage.createUser(getUser());
+        User user1 = getUser();
+        user1.setEmail("test1@mail.ru");
+        user1.setLogin("testLogin1");
+        User savedUser = userDbStorage.createUser(user1);
+        User user2 = getUser();
+        user2.setEmail("test2@mail.ru");
+        user2.setLogin("testLogin2");
+        User savedUser2 = userDbStorage.createUser(user2);
 
-        assertNotNull(followDbStorage.getCommonFriends(follow.getUserId(), common.getUserId())
-                , "Common friends is empty");
-    }*/
+        followDbStorage.addFriend(1L, 2L);
+        followDbStorage.addFriend(1L, 3L);
+        assertEquals(List.of(savedUser, savedUser2), followDbStorage.getAllFriends(1L));
+    }
+
+    @Test
+    void testGetEmptyFriendsOfFriend() {
+        userDbStorage.createUser(getUser());
+        User user = getUser();
+        user.setEmail("test1@mail.ru");
+        user.setLogin("testLogin1");
+        userDbStorage.createUser(user);
+        followDbStorage.addFriend(1L, 2L);
+        assertEquals(new ArrayList<>(), followDbStorage.getAllFriends(2L));
+    }
+
+    @Test
+    void testGetEmptyCommonFriends() {
+        assertEquals(new ArrayList<>(), followDbStorage.getCommonFriends(1L, 2L));
+    }
+
+    @Test
+    void testGetOneCommonFriend() {
+        userDbStorage.createUser(getUser());
+        User user1 = getUser();
+        user1.setEmail("test1@mail.ru");
+        user1.setLogin("testLogin1");
+        userDbStorage.createUser(user1);
+        User user2 = getUser();
+        user2.setEmail("test2@mail.ru");
+        user2.setLogin("testLogin2");
+        User savedUser2 = userDbStorage.createUser(user2);
+
+        followDbStorage.addFriend(1L, 2L);
+        followDbStorage.addFriend(1L, 3L);
+        followDbStorage.addFriend(2L, 3L);
+        assertEquals(List.of(savedUser2), followDbStorage.getCommonFriends(1L, 2L));
+    }
+
+    @Test
+    void testDeleteUserFriend() {
+        userDbStorage.createUser(getUser());
+        User user1 = getUser();
+        user1.setEmail("test1@mail.ru");
+        user1.setLogin("testLogin1");
+        userDbStorage.createUser(user1);
+
+        followDbStorage.addFriend(1L, 2L);
+        assertDoesNotThrow(() -> followDbStorage.deleteFriend(1L, 2L));
+    }
+
+    @Test
+    void testGetOneCommonFriendAfterDeletingOfFriend() {
+        userDbStorage.createUser(getUser());
+        User user1 = getUser();
+        user1.setEmail("test1@mail.ru");
+        user1.setLogin("testLogin1");
+        userDbStorage.createUser(user1);
+        User user2 = getUser();
+        user2.setEmail("test2@mail.ru");
+        user2.setLogin("testLogin2");
+        User savedUser2 = userDbStorage.createUser(user2);
+
+        followDbStorage.addFriend(1L, 2L);
+        followDbStorage.addFriend(1L, 3L);
+        followDbStorage.addFriend(2L, 3L);
+        followDbStorage.deleteFriend(1L, 2L);
+        assertEquals(List.of(savedUser2), followDbStorage.getCommonFriends(1L, 2L));
+    }
 }
